dependency updates + refactoring

+++ routes/api.php
@@ -19,31 +19,31 @@ Route::post('form-submissions/survey', 'Api\FormSubmissionController@submitSurve
 Route::post('form-submissions/sign-request', 'Api\FormSubmissionController@submitSignRequest');
 Route::post('form-submissions/volunteer', 'Api\FormSubmissionController@submitVolunteerRequest');
 Route::post('form-submissions/donation', 'Api\FormSubmissionController@submitDonationRequest');
-Route::post('tracking/site-view', 'Api\TrackingController@trackSiteView');
+Route::post('tracking/site-view', 'Api\Admin\Tracking\TrackingController@trackSiteView');
 
 Route::middleware(['client_credentials'])->group(function () {
 
-    Route::get('survey-submissions', 'Api\SurveySubmissionsController@index');
-    Route::get('survey-submissions/latest', 'Api\SurveySubmissionsController@getLatestSurveySubmissions');
-    Route::get('survey-submissions/{id}', 'Api\SurveySubmissionsController@retrieve');
-    Route::delete('survey-submissions/{id}', 'Api\SurveySubmissionsController@destroy');
+    Route::get('survey-submissions', 'Api\Admin\SurveySubmissions\SurveySubmissionsController@index');
+    Route::get('survey-submissions/latest', 'Api\Admin\SurveySubmissions\SurveySubmissionsController@getLatestSurveySubmissions');
+    Route::get('survey-submissions/{id}', 'Api\Admin\SurveySubmissions\SurveySubmissionsController@retrieve');
+    Route::delete('survey-submissions/{id}', 'Api\Admin\SurveySubmissions\SurveySubmissionsController@destroy');
 
-    Route::get('sign-requests', 'Api\SignRequestsController@index');
-    Route::get('sign-requests/latest', 'Api\SignRequestsController@getLatestSignRequests');
-    Route::get('sign-requests/{id}', 'Api\SignRequestsController@retrieve');
-    Route::delete('sign-requests/{id}', 'Api\SignRequestsController@destroy');
+    Route::get('sign-requests', 'Api\Admin\SignRequests\SignRequestsController@index');
+    Route::get('sign-requests/latest', 'Api\Admin\SignRequests\SignRequestsController@getLatestSignRequests');
+    Route::get('sign-requests/{id}', 'Api\Admin\SignRequests\SignRequestsController@retrieve');
+    Route::delete('sign-requests/{id}', 'Api\Admin\SignRequests\SignRequestsController@destroy');
 
-    Route::get('volunteers', 'Api\VolunteersController@index');
-    Route::get('volunteers/latest', 'Api\VolunteersController@getLatestVolunteers');
-    Route::get('volunteers/{id}', 'Api\VolunteersController@retrieve');
-    Route::delete('volunteers/{id}', 'Api\VolunteersController@destroy');
+    Route::get('volunteers', 'Api\Admin\Volunteers\VolunteersController@index');
+    Route::get('volunteers/latest', 'Api\Admin\Volunteers\VolunteersController@getLatestVolunteers');
+    Route::get('volunteers/{id}', 'Api\Admin\Volunteers\VolunteersController@retrieve');
+    Route::delete('volunteers/{id}', 'Api\Admin\Volunteers\VolunteersController@destroy');
 
-    Route::get('donations', 'Api\DonationsController@index');
-    Route::get('donations/latest', 'Api\DonationsController@getLatestDonations');
-    Route::get('donations/{id}', 'Api\DonationsController@retrieve');
-    Route::delete('donations/{id}', 'Api\DonationsController@destroy');
+    Route::get('donations', 'Api\Admin\Donations\DonationsController@index');
+    Route::get('donations/latest', 'Api\Admin\Donations\DonationsController@getLatestDonations');
+    Route::get('donations/{id}', 'Api\Admin\Donations\DonationsController@retrieve');
+    Route::delete('donations/{id}', 'Api\Admin\Donations\DonationsController@destroy');
 
-    Route::get('dashboard/get-stats', 'Api\DashboardController@getDashboardStats');
+    Route::get('dashboard/get-stats', 'Api\Admin\Dashboard\DashboardController@getDashboardStats');
 
 });
